[update] DiagramViewStateを別ファイルに切り分け
